+++ Inc/mmr_can.h
@@ -14,7 +14,7 @@
 #include "mmr_can_types.h"
 #include "mmr_can_optimize.h"
 #include "mmr_can_binary_literals.h"
-#include "mmr_can_scs_manager.h"
+#include "mmr_can_scs.h"
 
 #ifndef MMR_CAN_RX_FIFO
 #define MMR_CAN_RX_FIFO CAN_RX_FIFO0
@@ -181,6 +181,13 @@ extern MmrCanTickProvider __mmr_can_tickProvider;
  *  The function to use when fetching the current tick.
  */
 void MMR_CAN_SetTickProvider(MmrCanTickProvider tickProvider);
+
+/**
+ * @brief
+ * Returns the current tick by calling the
+ * configured tick provider.
+ */
+uint32_t MMR_CAN_GetCurrentTick();
 
 /**
  * @brief
